make fetch accept input of remoteType upstream, mine, or default
now that i'm not longer using git submodules for emacs packages I
don't automatically get the latest code of my fork when pulling on
the .emacs.d/ repo. So I have a need to fetch on more than just the
upstream now.

Making fetch fn more flexible to be able to fetch from the specified
remoteType.